% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_databases.R
\name{associated_genes_with_kegg}
\alias{associated_genes_with_kegg}
\title{Associate genes with KEGG}
\usage{
associated_genes_with_kegg(organism_code, classification = "module",
  genes = NULL)
}
\arguments{
\item{organism_code}{a three letter code}

\item{classification}{type of kegg classification to use: either module or pathway}

\item{optional}{character vector of gene names to return}
}
\description{
Find KEGG pathway IDs and names associated with an organisms' genes
}
\examples{
associated_genes_with_kegg("sce")
}

