% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cfs_to_hm3d}
\alias{cfs_to_hm3d}
\title{Convert flow units from cfs to hm3/day}
\usage{
cfs_to_hm3d(x)
}
\arguments{
\item{x}{vector of flows in cfs}
}
\value{
vector of flows in hm3/day
}
\description{
Convert flow units from cfs to hm3/day
}
\examples{
cfs_to_hm3d(c(1.2, 3.4, 12.3))
}

