% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{convert_metabolomics_to_wide_output}
\alias{convert_metabolomics_to_wide_output}
\title{Convert Metabolomics to Wide Output}
\usage{
convert_metabolomics_to_wide_output(normalized_data, metabolite_info, output)
}
\arguments{
\item{normalized_data}{Normalized metabolomics quantitative data}

\item{metabolite_info}{Information specific to each measured metabolite}

\item{output}{Path to output}
}
\value{
writes a tsv to output
}
\description{
Convert normalized metabolomics data to wide-data and save to output.
}
\examples{
directory = system.file("extdata", "maven_axon.xlsx", package = "fluxr")
maven_data <- read_MAVEN_data(directory)
normalized_data <- normalize_metabolomics(maven_data$metabolomics_data)
metabolite_info <- maven_data$metabolite_info
convert_metabolomics_to_wide_output(output = "fluxr_trial_output.txt")
}

