% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_databases.R
\name{download_BRENDA_regulators}
\alias{download_BRENDA_regulators}
\title{Download BRENDA Regulators}
\usage{
download_BRENDA_regulators()
}
\description{
Calls python to summarize all regulators in BRENDA by E.C. number.
}

