% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_databases.R
\name{extended_gene_annotation}
\alias{extended_gene_annotation}
\title{Extended Gene Annotation}
\usage{
extended_gene_annotation(dataset)
}
\arguments{
\item{dataset}{a biomaRt dataset contained within listDatasets(useMart('ensembl'))}
}
\value{
a tibble combining multiple IDs for a gene
}
\description{
Extended Gene Annotation
}
\examples{
extended_gene_annotation("scerevisiae_gene_ensembl")
}

