% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux_fwm.R
\name{flux_fwm}
\alias{flux_fwm}
\title{Estimate daily loads using flow-weighted mean concentration by year}
\usage{
flux_fwm(df, n_min = 5)
}
\arguments{
\item{df}{data frame containing observed concentrations, flows, and loads}

\item{n_min}{minimum number of samples to use fwm concentration by water year}
}
\value{
input data frame with estimated concentration and loads
}
\description{
Estimate daily loads using flow-weighted mean concentration by year
}

