% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flux_fwm_strat}
\alias{flux_fwm_strat}
\title{Estimate daily loads using flow-weighted mean concentration by water year and stratified flow}
\usage{
flux_fwm_strat(df, n_min = 5)
}
\arguments{
\item{df}{data frame containing observed concentrations, flows, and loads}

\item{n_min}{minimum number of samples to use fwm concentration by water year}
}
\value{
input data frame with estimated concentration and loads
}
\description{
Estimate daily loads using flow-weighted mean concentration by water year and stratified flow
}

