% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxR.R
\name{flux_meld}
\alias{flux_meld}
\title{Flux meld
#'Rbinds and left_joins all co2 and ch4 datasets together in the given directory,
adds NA to esxcluded flux. Output is saved to "flux_melded.csv"}
\usage{
flux_meld(directory)
}
\arguments{
\item{directory}{The path as a character string to the given directory}
}
\description{
Flux meld
#'Rbinds and left_joins all co2 and ch4 datasets together in the given directory,
adds NA to esxcluded flux. Output is saved to "flux_melded.csv"
}
\examples{
flux_meld(C:/path/to/files)
}
