% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux_regression.R
\name{flux_regression}
\alias{flux_regression}
\title{Estimate daily loads using multi-variate regression}
\usage{
flux_regression(x, interp = FALSE, max_interval = 90,
  extrapolate_flow = FALSE, start_month = 10, predict_wyear_range = NULL)
}
\arguments{
\item{x}{data frame containing observed concentrations, flows, and loads}

\item{interp}{flag to interpolate residuals}

\item{max_interval}{maximum interval for using interpolation}

\item{extrapolate_flow}{flag to allow extrapolation of flow variables for prediction (Qderiv, Q)}

\item{start_month}{first month of water year}

\item{predict_wyear_range}{vector of length 2 specifying range of water years for final predications (or NULL to use span of data)}
}
\value{
list
}
\description{
Estimate daily loads using multi-variate regression
}

