% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux_stats.R
\name{flux_stats}
\alias{flux_stats}
\title{Compute error statistics of load estimates}
\usage{
flux_stats(x, model, n.coeff, method)
}
\arguments{
\item{x}{data frame containing observed, predicted and residual loads}

\item{model}{linear model object of type lm}

\item{n.coeff}{number of coefficients used to make predictions (for computing degrees of freedom)}

\item{method}{name of method for labelling results}
}
\value{
list of error statistics
}
\description{
Compute error statistics of load estimates
}

