% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxr-package.R
\docType{package}
\name{fluxr-package}
\alias{fluxr-package}
\alias{fluxr}
\title{The 'fluxr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: Michelle Kelly \email{mckelly1@mtu.edu} (\href{https://orcid.org/0000-0003-0123-2527}{ORCID})

}
