% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mcmc_mat.R
\name{format_mcmc_mat}
\alias{format_mcmc_mat}
\title{\code{format_mcmc_mat} Format MCMC output into a one-row data.frame}
\usage{
format_mcmc_mat(mcmc_mat, keep_mcmc, runmcmc_out, data_list)
}
\arguments{
\item{mcmc_mat}{matrix as extracted from Stan}

\item{names_params}{character vector of the names of the parameters}

\item{names_stats}{character vector of the names of the statistics}
}
\description{
For split_dates models. Formats output into a one-row data.frame for
row-binding with other such data.frames
}
\keyword{internal}
