% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mcmc_mat_twostation.R
\name{format_mcmc_mat_twostation}
\alias{format_mcmc_mat_twostation}
\title{\code{format_mcmc_mat_twostation} Format MCMC output into a one-row data.frame}
\usage{
format_mcmc_mat_twostation(mcmc_mat, keep_mcmc, runmcmc_out)
}
\arguments{
\item{mcmc_mat}{matrix as extracted from Stan}

\item{names_params}{character vector of the names of the parameters}

\item{names_stats}{character vector of the names of the statistics}
}
\description{
\code{format_mcmc_mat_twostation} Format MCMC output into a one-row data.frame
}
\keyword{internal}
