% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odensity.R
\name{h2odensity}
\alias{h2odensity}
\title{Calculate water density at temperatures between 0 and 40 degC}
\usage{
h2odensity(temp)
}
\arguments{
\item{temp}{Water temperature in degrees C}
}
\value{
A numeric vector in units of kg/m3
}
\description{
Calculate water density at temperatures between 0 and 40 degC
}
\references{
Tanaka, M., Girard, G., Davis, R., Peuto, A., & Bignell, N. (2001). Recommended table for the density of water between 0 C and 40 C based on recent experimental reports. Metrologia, 38(4), 301–309. https://doi.org/10.1088/0026-1394/38/4/3
}
