% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interp_range.R
\name{interp_range}
\alias{interp_range}
\title{Interpolate missing values with fixed ranges around discrete samples}
\usage{
interp_range(x, y, max_interval, fill = 0)
}
\arguments{
\item{x}{a vector of timesteps (numeric or date-like)}

\item{y}{a vector of discrete values with NAs indicating missing values}

\item{max_interval}{the maximum number of timesteps for using linear interpolation}

\item{fill}{value to fill beginning and ending values outside max_interval}
}
\value{
vector of interpolated values corresponding to x
}
\description{
Interpolate missing values with fixed ranges around discrete samples
}
\examples{
interp_range(x=seq(1,13), y=c(NA, NA, 1, NA, NA, 2, NA, NA, NA, 3, NA, NA, NA), max_interval=2, fill=0)
}

