% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcor_N2.R
\name{Kcor_N2}
\alias{Kcor_N2}
\title{\code{Kcor_N2} Estimate kN2, given water temperature and K600}
\usage{
Kcor_N2(temp, K600, n = 0.5)
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{K600}{Normalized K600 reaeration rate}

\item{n}{Default is 0.5 for wavy water, but 0.667 refers to still water, see Jahne 1987}
}
\value{
Numeric vector
}
\description{
Esimate k values (not temperature or gas normalized) from K600 (temperature and gas normalized). See Wanninkhof 1992 and Hall et al 2016 supplemental material.
}
\references{
Raymond, P. A., Zappa, C. J., Butman, D., Bott, T. L., Potter, J., Mulholland, P., et al. (2012). Scaling the gas transfer velocity and hydraulic geometry in streams and small rivers: Gas transfer velocity and hydraulic geometry. Limnology and Oceanography: Fluids and Environments, 2(1), 41–53. https://doi.org/10.1215/21573689-1597669
}
