% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxR.R
\name{lgr_flux}
\alias{lgr_flux}
\title{Flux Calculator}
\usage{
lgr_flux(
  lgr_data,
  start_times,
  site_name,
  chamber = T,
  chamb_volume,
  chamb_area
)
}
\arguments{
\item{lgr_data}{character,  path to melded or unmelded LGR data including file extension}

\item{start_times}{character, path to file with start times}

\item{site_name}{character, string of site name}

\item{chamber}{boolean, if true function will grab chamber type from start_times, if
false function will use chamb_volume and chamb_area}

\item{chamb_volume}{numerical, volume of chamber m3}

\item{chamb_area}{numerical, cross section area of chamber m2}
}
\description{
Takes in lgr melded or unmelded data, as csv or txt. IF TXT FIRST ROW WILL BE SKIPPED.
Runs and plots initial regressions, asks for inputs for editing,
reruns and checks regressions and saves to output directories.
}
\details{
Directory structure is important. Working directory with Site Folders inside
with Date Folders inside of that.

Must be run on one site at a time.
}
\examples{
lgr_flux(
lgr_data = "c:/path/to/lgr.csv",
start_times = "c:/path/to/start_times.csv,
site_name = "Site_name",
chamber = T)

lgr_flux(
lgr_data = "c:/path/to/lgr.csv",
start_times = "c:/path/to/start_times.csv,
site_name = "Site_name",
chamber = F,
chamb_volume = 0.02
chamb_area = 0.02)


}
