% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxR.R
\name{lgr_meld}
\alias{lgr_meld}
\title{Dataset binder}
\usage{
lgr_meld(directory)
}
\arguments{
\item{directory}{The path as a character string to the given directory}
}
\description{
Rbinds all LGR datasets together in the given directory,
removes unnecessary rows and deletes empty files. Output is saved to "melded.csv"
}
\examples{
lgr_meld(C:/path/to/files)
}
