% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_stats}
\alias{load_stats}
\title{Compute error statistics of load estimates}
\usage{
load_stats(df, n.coeff, method)
}
\arguments{
\item{df}{data frame containing observed, predicted and residual loads}

\item{n.coeff}{number of coefficients used to make predictions (for computing degrees of freedom)}

\item{method}{name of method for labelling results}
}
\value{
list of error statistics
}
\description{
Compute error statistics of load estimates
}

