% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_breaks.R
\name{log_breaks}
\alias{log_breaks}
\title{Logarithmic breaks}
\usage{
log_breaks(x, y)
}
\arguments{
\item{x}{vector of first integers}

\item{y}{vector of magnitudes}
}
\description{
Logarithmic breaks
}
\examples{
log_breaks(seq(1, 9), c(10, 100, 1000))
}

