% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{normalize_metabolomics}
\alias{normalize_metabolomics}
\title{Normalize Metabolomics}
\usage{
normalize_metabolomics(metabolomics_data)
}
\arguments{
\item{metabolomics_data}{a tidy data.frame generated from \code{\link{read_MAVEN_data}}}
}
\value{
metabolomics_data with an added column containing normalized ion counts
}
\description{
Using a robust metabolomics normalization rescale ion counts based on the consensus signal of each sample.
}
\details{
This normalization was reported in Kamphorst et al. 2015, see \url{https://github.com/shackett/Pancreatic_tumor_metabolomics}.
}
\examples{
directory = system.file("extdata", "maven_axon.xlsx", package = "fluxr")
normalize_metabolomics(read_MAVEN_data(directory)$metabolomics_data)
}

