% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_flux_monthly.R
\name{plot_flux_monthly}
\alias{plot_flux_monthly}
\title{Plot monthly summary of flux load estimation model}
\usage{
plot_flux_monthly(loads, site, variable, title = NULL)
}
\arguments{
\item{loads}{output of flux estimation function}

\item{site}{name of site}

\item{variable}{name of variable}

\item{title}{plot title (if NULL, uses a default)}
}
\description{
Plot monthly summary of flux load estimation model
}

