% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata_bayes_twostation.R
\name{prepdata_bayes_twostation}
\alias{prepdata_bayes_twostation}
\title{\code{prepdata_bayes_twostation} Prepare two-station data for Bayesian modeling}
\usage{
prepdata_bayes_twostation(data, specs, up.name, down.name)
}
\arguments{
\item{data}{Dataset containing columns: \code{solar.time}, \code{location}, \code{DO.obs}, \code{DO.sat}, \code{temp.water}, \code{light}, \code{depth}, \code{tt}, \code{pressure}}

\item{specs}{List returned by \code{set_specs()}}

\item{ply_date}{Date to model, YYYY-MM-DD}
}
\description{
\code{prepdata_bayes_twostation} Prepare two-station data for Bayesian modeling
}
