% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_DO.R
\name{return_DO}
\alias{return_DO}
\title{\code{return_DO} Return a data.frame containing solar time, observed DO, and modeled DO}
\usage{
return_DO(data_list, model_output)
}
\arguments{
\item{data_list}{model input data, as formatted by \code{prepdata_bayes_twostation()}}

\item{model_output}{model output, as returned by \code{runstan_bayes_twostation()}}
}
\description{
\code{return_DO} Return a data.frame containing solar time, observed DO, and modeled DO
}
\keyword{internal}
