% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_gas.R
\name{return_gas}
\alias{return_gas}
\title{\code{return_gas} Return a data.frame containing solar time, observed gas, and modeled gas}
\usage{
return_gas(modname, instresults, inputdata)
}
\description{
\code{return_gas} Return a data.frame containing solar time, observed gas, and modeled gas
}
\keyword{internal}
