% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_twostation.R
\name{run_twostation}
\alias{run_twostation}
\title{\code{run_twostation} Fit a two-station Bayesian metabolism model for O2 or N2 flux}
\usage{
run_twostation(data, specs, upname = "upstream", downname = "downstream")
}
\arguments{
\item{data}{dataframe containing columns \code{solar.time}, \code{location}, \code{DO.obs}, \code{DO.sat}, \code{temp.water}, \code{light}, \code{depth}, \code{tt}, \code{pressure}}

\item{specs}{list of model specifications as returned by \code{set_twostation_specs()}}

\item{upname}{a character string specifying the name of the upstream sampling station in the \code{location} column. Defaults to "upstream"}

\item{downname}{a character string specifying the name of the downstream sampling station in the \code{location} column. Defaults to "downstream"}
}
\value{
Numeric vector
}
\description{
\code{run_twostation} Fit a two-station Bayesian metabolism model for O2 or N2 flux
}
\references{
Raymond, P. A., Zappa, C. J., Butman, D., Bott, T. L., Potter, J., Mulholland, P., et al. (2012). Scaling the gas transfer velocity and hydraulic geometry in streams and small rivers: Gas transfer velocity and hydraulic geometry. Limnology and Oceanography: Fluids and Environments, 2(1), 41–53. https://doi.org/10.1215/21573689-1597669
}
