% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan_bayes_twostation.R
\name{runstan_bayes_twostation}
\alias{runstan_bayes_twostation}
\title{\code{runstan_bayes_twostation}  Fit two-station Bayesian metabolism model}
\usage{
runstan_bayes_twostation(data_list, specs)
}
\arguments{
\item{data_list}{list of inputs to Stan model, as returned by \code{prepdata_bayes_twostation()}}

\item{specs}{list of model specifications, as returned by \code{set_specs()}}
}
\description{
\code{runstan_bayes_twostation}  Fit two-station Bayesian metabolism model
}
