% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_twostation_specs.R
\name{set_twostation_specs}
\alias{set_twostation_specs}
\title{\code{set_twostation_specs} Generate a list of model specifications}
\usage{
set_twostation_specs(
  model_name,
  keep_mcmc = TRUE,
  keep_mcmc_data = TRUE,
  GPP_daily_mu = 3.1,
  GPP_daily_lower = -Inf,
  GPP_daily_sigma = 6,
  ER_daily_mu = -7.1,
  ER_daily_upper = Inf,
  ER_daily_sigma = 7.1,
  N2consume_daily_mu = -1,
  N2consume_daily_lower = -Inf,
  N2consume_daily_sigma = 5,
  DN_daily_mu = 1,
  DN_daily_upper = Inf,
  DN_daily_sigma = 5,
  K600_daily_meanlog = log(12),
  K600_daily_sdlog = 0.05,
  err_obs_iid_sigma_scale = 0.03,
  n_chains = 4,
  n_cores = 4,
  burnin_steps = 500,
  saved_steps = 500,
  thin_steps = 1,
  verbose = FALSE
)
}
\description{
This function is based on the function of the same name from
\code{streamMetabolizer}, just simplified to only parameters relavant for
our model type (Bayes, observation error, reaeration depends on modeled gas
at downstream station).
}
\details{
Default values for GPP, ER, and K600 are the same as \code{streamMetabolizer}.
Default values for N2consume and DN are the same as those in Nifong et al.
2020.
}
