% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-williamson.R
\docType{data}
\name{williamson}
\alias{williamson}
\title{Williamson River flow and water quality data}
\format{A \code{\link{data.frame}} with 7,671 observations.}
\usage{
williamson
}
\description{
Approximately biweekly water quality data
}
\section{Dimensions}{


\itemize{
  \item \code{DATE}: date
}
}

\section{Measures}{


\itemize{
  \item \code{Q}: daily streamflow (hm3/d)
  \item \code{C}: sampled concentration (ppb)
}
}
\examples{
williamson
}

