% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wyear}
\alias{wyear}
\title{Get water year component of a date-time.}
\usage{
wyear(x, start_month = 10)
}
\arguments{
\item{x}{a date-time object}

\item{start_month}{starting month of the water year}
}
\value{
the water year elements of x as a number
}
\description{
Get water year component of a date-time.
}
\examples{
wyear(lubridate::ymd('2000-09-30') + lubridate::days(c(0,1,2)), start_month=10)
}

