% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux_fwm_strat.R
\name{flux_fwm_strat}
\alias{flux_fwm_strat}
\title{Estimate daily loads using flow-weighted mean concentration by water year and stratified flow}
\usage{
flux_fwm_strat(df, n_min = 5)
}
\arguments{
\item{df}{data frame containing observed concentrations, flows, and loads}

\item{n_min}{minimum number of samples to use fwm concentration by water year}
}
\value{
data frame \code{df} with estimated daily concentration and loads
}
\description{
Estimate daily loads using flow-weighted mean concentration by water year and stratified flow
}

