% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flux.R
\name{flux_regression}
\alias{flux_regression}
\title{Estimate daily loads using multi-variate regression}
\usage{
flux_regression(df, interp = FALSE, max_interval = 90)
}
\arguments{
\item{df}{data frame containing observed concentrations, flows, and loads}

\item{interp}{flag to use interpolated residuals}

\item{max_interval}{maximum interval for using interpolation}
}
\value{
input data frame with estimated concentration and loads
}
\description{
Estimate daily loads using multi-variate regression
}

